# docs/swagger.yaml
...
basePath: /api/v1
definitions:
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ObservationPoint:
    properties:
      activeObservers:
        type: integer
      hour:
        description: epoch ms, start of the 1-hour bucket
        type: integer
      iata:
        type: string
      observationCount:
        type: integer
      uniquePackets:
        type: integer
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.PayloadBreakdownItem:
    properties:
      count:
        type: integer
      payloadType:
        type: integer
      payloadTypeName:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.RadioPreset:
    properties:
      count:
        description: number of observers or nodes on this preset in this IATA
        type: integer
      iata:
        type: string
      preset:
        description: '"freqMhz,bwKhz,sf" e.g. "910.525,62.5,7"'
        type: string
      sourceType:
        description: '"observer" or "node"'
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop:
    properties:
      confidence:
        description: '"high", "ambiguous", or "none"'
        type: string
      nodes:
        description: empty for "none", one for "high", multiple for "ambiguous"
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode'
        type: array
      snr:
        type: number
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode:
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      publicKey:
        description: hex-encoded prefix used for resolution
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeStats:
    properties:
      name:
        description: normalized scope name e.g. "#bc"
        type: string
      nodeCount:
        description: distinct nodes with this as their default scope
        type: integer
      observerCount:
        description: distinct observers that forwarded packets in this scope
        type: integer
      packetCount:
        description: distinct packets matched to this scope
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.StatsOverview:
    properties:
      activeIatas:
        type: integer
      activeObservers:
        type: integer
      totalObservations:
        type: integer
      totalPackets:
        type: integer
      windowHours:
        description: always 24 for now
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TopNode:
    properties:
      iata:
        type: string
      lastHeard:
        description: epoch ms
        type: integer
      nodeId:
        type: string
      nodeName:
        type: string
      nodeType:
        type: integer
      nodeTypeName:
        type: string
      observationCount:
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TopObserver:
    properties:
      displayName:
        type: string
      iata:
        type: string
      observationCount:
        type: integer
      observerId:
        type: string
      observerType:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TraceDetail:
    properties:
      packets:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TracePacket'
        type: array
      traceTag:
        description: hex-encoded 4-byte tag
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TracePacket:
    properties:
      firstHeardAt:
        description: epoch ms
        type: integer
      lastHeardAt:
        description: epoch ms
        type: integer
      packetHash:
        type: string
      resolvedRoute:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop'
        type: array
      routeType:
        type: integer
      routeTypeName:
        type: string
      scope:
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.TraceTagSummary:
    properties:
      firstHeardAt:
        description: epoch ms
        type: integer
      iataCount:
        description: number of distinct IATAs where heard
        type: integer
      lastHeardAt:
        description: epoch ms
        type: integer
      packetCount:
        description: number of packets with this trace tag
        type: integer
      traceTag:
        description: hex-encoded 4-byte tag
        type: string
    type: object
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: MIT
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: "1.0"
paths:
...
  /scopes:
    get:
      parameters:
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List transport scopes
      tags:
      - Scopes
...
  /stats/observations:
    get:
      parameters:
      - description: Filter by IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Start of window epoch ms (default 7 days ago)
        in: query
        name: since
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObservationPoint'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Hourly observation time series
      tags:
      - Stats
  /stats/overview:
    get:
      parameters:
      - description: Filter by IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsOverview'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Network overview stats (last 24h)
      tags:
      - Stats
  /stats/payload-breakdown:
    get:
      parameters:
      - description: Filter by IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Start of window epoch ms (default last 24h)
        in: query
        name: since
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PayloadBreakdownItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Observation counts by payload type (last 24h by default)
      tags:
      - Stats
  /stats/radio-presets:
    get:
      parameters:
      - description: Filter by preset string e.g. 910.525,62.5,7
        in: query
        name: preset
        type: string
      - description: Filter by IATA code
        in: query
        name: iata
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RadioPreset'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Radio preset usage by IATA
      tags:
      - Stats
  /stats/scopes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Scope statistics
      tags:
      - Stats
  /stats/top-nodes:
    get:
      parameters:
      - description: Filter by exact IATA code (case-sensitive)
        in: query
        name: iata
        type: string
      - description: Max results (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TopNode'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Top N nodes by observation count (from materialized view)
      tags:
      - Stats
  /stats/top-observers:
    get:
      parameters:
      - description: Filter by IATA code (case-insensitive)
        in: query
        name: iata
        type: string
      - description: Start of window epoch ms (default last 24h)
        in: query
        name: since
        type: integer
      - description: Max results (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TopObserver'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Top N observers by observation count (last 24h by default)
      tags:
      - Stats
  /traces:
    get:
      parameters:
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      - description: Filter by transport scope name
        in: query
        name: scope
        type: string
      - description: Filter by first_heard_at >= since (epoch ms)
        in: query
        name: since
        type: integer
      - description: Filter by first_heard_at <= until (epoch ms)
        in: query
        name: until
        type: integer
      - description: last_heard_at epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TraceTagSummary'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List trace tags
      tags:
      - Traces
  /traces/{tag}:
    get:
      parameters:
      - description: Trace tag hex e.g. a3f1b2c4
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TraceDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get full trace detail by tag
      tags:
      - Traces
schemes:
- http
- https
swagger: "2.0"
...